backend/requirements.txt
fastapi
uvicorn[standard]
firebase-admin
paho-mqtt
influxdb-client
apscheduler
pandas
numpy
scikit-learn
xgboost
joblib
python-dotenv
pydantic

